Move the failure dump's rendering out of _DemoBase, and grade it
The narrow slice of #147, re-cut against the numbers rather than the ones the
issue was filed with. The re-measurement is on the issue; the short version is
that #138 closed most of the leak it describes — `_DemoBase` is 1,700 lines not
2,408, `Recorder` now calls *no* base private method, and two of the six
proposed collaborators no longer exist.

What the evidence still pointed at was the failure dump, and specifically the
half of it that is a function of documents rather than of a live recorder.

## What moved, and why these five

Chosen by how much recorder state each method touches, not by name:

| method | lines | `self.*` used |
|---|---:|---:|
| `_failure_md` | 76 | **0** |
| `_failure_summary` | 14 | 1 |
| `_failed_beat` | 13 | 1 |
| `_clear_failure_dir` | 34 | 1 |
| `_clear_failure_marker` | 22 | 1 |

`_failure_md` was a 76-line pure function written as a method. `failure.py`
already owned `FAILURE_DIR`, `FAILURE_MARKER` and `FAILURE_SCHEMA`, so the
rendering half had a home waiting for it.

`_DemoBase`: 1,700 -> 1,536 lines, 56 -> 51 methods.

**`_failure_screen` is deliberately not in that list**, though the issue's
sketch implies it. It returns `None` and both subclasses override it — it is the
seam, not the rendering. `_failure_doc`, `_write_failure`, `_write_failure_marker`,
`_build_failure` and `_report_failure` stay too: they read `_converted`,
`_media_path`, `_beats`, `_issues` and `out_dir`, and moving them is a refactor
with real risk rather than a code move.

## The move is proven, not asserted

Old and new ASTs compared statement by statement, with only the four
substitutions the move is *allowed* to make applied to the original:
`self.out_dir` -> `out_dir`, `self._beats` -> `beats`, `self._failed_beat()` ->
`failed_beat(beats)`, and `reversed(beats)` -> `reversed(list(beats))`. All five
bodies come back identical.

## markdown.py

`render_failure_md` needs `_md_cell`, which lived in `timeline.py` — but
`timeline` imports `failure` for its paths, so importing back was a cycle. The
two markdown helpers move to a new leaf module that imports nothing from the
package.

Three modules render markdown (`timeline`, `frames`, `failure`) and `_md_cell`
is the only thing between an exception message holding a pipe and a table
gaining a column. A second copy is a second place to forget the backslash, so
they are shared rather than duplicated.

## What is graded now

18 tests, and the first this artifact has ever had. `tests/smoke`'s four crash
takes grade that `failure.md` *exists* and holds a marker string; nothing graded
what it says. That matters for this file more than most: it is read at the moment
somebody's demo just broke, and a dump that blames the wrong beat or claims a
previous run's mp4 is this take's is the artifact-lies failure this repo treats
as blocking — on the one path where the reader has no working recording to check
it against.

Both directions of the #46 sentence are graded, which is the point of the pair:
a take that encoded nothing disowns the mp4 beside it and names the marker, and
a take that did encode claims its own recording. Either alone passes on a
renderer stuck in one mode.

## Injected

    the dump blames the last beat, not the one that raised
    an exception message reaches the dump uncapped
    a take that encoded no mp4 stops disowning the one beside it (#46)
    the dump stops saying how many issues it left out
    clearing a previous run's dump takes the whole directory with it

Two pre-existing entries had to be retargeted from `timeline.py` at
`markdown.py` — the exactly-once guard refused to proceed on both, which is what
it is for. 34 of 34 injections caught.

One of my own tests was too weak and the manifest said so: the summary test had
a single beat, so "the last beat" and "the beat that raised" were the same row
and the `beats[-1]` injection stayed green. It now records a beat *after* the
failing one, which is what a take does when it logs a teardown beat.

Full `tests/smoke` passes, 98 reported checks. `tests/unit` is 73 tests.

Refs #147
