+++ skills/demo-video/helpers/demo_recording/failure.py
@@ -1,12 +1,28 @@
 """Where a take that did not finish writes itself down.
 
-Only the constants live here — the paths, the schema and the message budget.
-The documents themselves are built by the recorder, because building them
-means reading a live page, and they are built in memory and refused whole if
-the mask cannot be vouched for.
+The constants — paths, schema, message budget — and the half of the dump that is
+a function of documents rather than of a live recorder: rendering `failure.md`,
+naming the beat that raised, and taking a previous run's dump off disk.
+
+**What is not here is what needs the recorder.** `_failure_doc` reads the
+buffered page text, the issue log and the media path off `self`;
+`_failure_screen` is the hook a medium overrides to say what its screen holds;
+`_write_failure` needs to know whether the encode happened. Those stay in
+`core`, and the split is drawn exactly where the state stops: everything below
+takes its inputs as arguments and can be checked without a browser (#147).
+
+The docstring used to say the documents "are built in memory and refused whole
+if the mask cannot be vouched for". There is no mask (#138); a document is built
+and written.
 """
 
 from __future__ import annotations
+
+import sys
+from collections.abc import Sequence
+from pathlib import Path
+
+from .markdown import _fmt_t, _md_cell
 
 # -- failure artifacts (issues #11, #20, #24, #32, #46) ----------------------
 #
@@ -67,3 +83,178 @@ FAILURE_MARKER = "demo-video-FAILED.md"
 # per-beat `error.message` is **not** capped — it is the machine-readable copy
 # and something may want to match on it — so nothing is lost by capping here.
 FAILURE_MESSAGE_CHARS = 2_000
+
+
+def failed_beat(beats: Sequence[dict]) -> dict | None:
+    """The beat whose verb raised, or None if the failure was between beats.
+
+    Reads `error`, which `_beat` stamps (issue #24), rather than assuming
+    the last beat is the culprit — a storyboard can raise in its own code
+    between two verbs, and blaming the last beat that *worked* is exactly
+    the confidently-wrong attribution `_attributed_beat` refuses to make
+    for issues.
+    """
+    for beat in reversed(list(beats)):
+        if "error" in beat:
+            return beat
+    return None
+
+
+def failure_summary(
+    exc_type: type, exc: BaseException | None, beats: Sequence[dict]
+) -> dict:
+    """What came out of the `with`, and which beat it came out of.
+
+    Unscrubbed: every consumer (`_timeline_doc`, the dump, the marker)
+    masks it on the way to its own file, and each has a different mask.
+    """
+    beat = failed_beat(beats)
+    message = str(exc) if exc is not None else ""
+    return {
+        "type": exc_type.__name__,
+        "message": message[:FAILURE_MESSAGE_CHARS],
+        "beat": None if beat is None else beat.get("index"),
+        "verb": None if beat is None else beat.get("verb"),
+    }
+
+
+def render_failure_md(doc: dict) -> str:
+    """The human half. Pure function of the document above, so it inherits
+    its masking rather than re-deriving it."""
+    failure = doc.get("failure") or {}
+    beat = doc.get("beat") or {}
+    where = (
+        f"beat {failure.get('beat')} (`{_md_cell(failure.get('verb'))}`"
+        + (
+            f", target `{_md_cell(beat.get('selector'))}`"
+            if beat.get("selector")
+            else ""
+        )
+        + ")"
+        if failure.get("beat") is not None
+        else "between beats — no verb was running when it happened"
+    )
+    out = [
+        "# This take did not finish",
+        "",
+        f"`{doc.get('recorder')}` · {doc.get('when')} · "
+        f"{doc.get('beats_recorded')} beats recorded",
+        "",
+        f"**{_md_cell(failure.get('type'))}** at {where}.",
+        "",
+        f"> {_md_cell(failure.get('message')) or '(no message)'}",
+        "",
+        "## What is here",
+        "",
+        "| file | what it is |",
+        "|---|---|",
+        "| `failure.json` | this, machine-readable: the failing beat in "
+        "full, every issue the take recorded, and what was written |",
+    ]
+    if doc.get("media_written_by_this_take"):
+        out.append(
+            f"| `last-frame.png` | the final frame of "
+            f"`{doc.get('media')}` — what was on screen when it stopped |"
+        )
+    if doc.get("screen_captured"):
+        out.append(
+            "| `screen.txt` | the page's accessibility tree (web) or the "
+            "rendered terminal buffer, read at the end of the take |"
+        )
+    out += ["", "## What the app said", ""]
+    issues = doc.get("issues") or []
+    if not issues:
+        out.append(
+            "Nothing. No console errors, failed requests or non-zero exits "
+            "were recorded, so the app did not announce this."
+        )
+    else:
+        for issue in issues:
+            seat = (
+                "before the first beat"
+                if issue.get("beat") is None
+                else f"beat {issue['beat']} (`{_md_cell(issue.get('verb'))}`)"
+            )
+            out.append(
+                f"- **{_md_cell(issue.get('kind'))}** — {seat} at "
+                f"{_fmt_t(issue.get('t'))}s: {_md_cell(issue.get('message'))}"
+            )
+        total = doc.get("issue_count", len(issues))
+        if total > len(issues):
+            out.append(f"- …and {total - len(issues)} more, not recorded.")
+    out += [
+        "",
+        "## The recording",
+        "",
+        (
+            f"`{doc.get('media')}` beside this folder is **this** take's "
+            f"partial recording — the webm the browser had in hand when the "
+            f"storyboard gave up, converted rather than discarded."
+            if doc.get("media_written_by_this_take")
+            else f"**This take encoded no mp4.** Any `{doc.get('media')}` in "
+            f"the folder above is a *previous* run's and is not a recording "
+            f"of this failure — see `{FAILURE_MARKER}`."
+        ),
+        "",
+    ]
+    return "\n".join(out).rstrip() + "\n"
+
+
+def clear_failure_dir(out_dir: Path) -> list[str]:
+    """Take a previous run's `failure/` off disk. Returns what went.
+
+    Same reasoning as `_clear_stale_evidence`, and the same hazard: this
+    directory holds a text dump of the page — an ARIA tree or a terminal
+    buffer — so a stale one sitting beside a *fresh* take is both a lie
+    about which run failed and a file that may hold the very value this
+    take was rewritten to hide. Bounded to the names `_write_failure`
+    writes, never the directory, so nothing anybody put here is touched.
+    """
+    directory = out_dir / FAILURE_DIR
+    gone: list[str] = []
+    if not directory.is_dir():
+        return gone
+    for name in ("failure.json", "failure.md", "screen.txt", "last-frame.png"):
+        path = directory / name
+        try:
+            if path.is_file():
+                path.unlink()
+                gone.append(f"{FAILURE_DIR}/{name}")
+        except OSError:  # noqa: PERF203 - report what could not be removed
+            print(
+                f"demo-video: WARNING — could not delete {path}, which is "
+                f"a previous take's failure dump and describes a run "
+                f"this one is not",
+                file=sys.stderr,
+            )
+    try:
+        next(directory.iterdir())
+    except StopIteration:
+        directory.rmdir()
+    except OSError:
+        pass
+    return gone
+
+
+def clear_failure_marker(out_dir: Path) -> None:
+    """Take a previous run's marker off disk once a take succeeds.
+
+    The other half of #46 and not an optional one: a marker left beside a
+    freshly-written demo.mp4 is the same lie inverted, and it is the one
+    that makes people stop believing the marker at all.
+    """
+    marker = out_dir / FAILURE_MARKER
+    try:
+        if marker.is_file():
+            marker.unlink()
+            print(
+                f"demo-video: this take wrote its own artifacts, so the "
+                f"{FAILURE_MARKER} a previous run left here is gone",
+                file=sys.stderr,
+            )
+    except OSError as exc:
+        print(
+            f"demo-video: WARNING — could not delete {marker} ({exc}). It "
+            f"describes a previous run, not this one.",
+            file=sys.stderr,
+        )
